Supermarket Management System
This program can keep track of all management and billing activities in a supermarket. It holds a record of all the customers, employees, and products. It can be used to add, see details of multiple entries, and delete and update customers, employees, and products. It can also be used for creating, editing, viewing and checking out the cart by a customer with a bill containing complete customer details about the customer, date, time, and payment method for the purchase.